Реализация интерфейса вынесена в отдельный файл
