+++ server_friends/tests/mock_classes.h
@@ -7,7 +7,7 @@
 
 #include <gtest/gtest.h>
 #include <gmock/gmock.h>
-#include <server_interface.h>
+#include "server_interface.h"
 #include <memory>
 #include "utilty.h"
 #include "friends_data_base.h"
